fix(build): circular types tsc errors
